Move and rename packages
