+++ src/main/java/com/pathdb/storage/NodeSize.java
@@ -7,7 +7,7 @@
 
 package com.pathdb.storage;
 
-import com.pathdb.pathIndex.tree.IndexTree;
+import com.pathdb.pathIndex.diskBased.IndexTree;
 
 import java.io.IOException;
