+++ src/test/java/com/pathdb/pathIndex/TreeNodeIDManagerTest.java
@@ -7,8 +7,8 @@
 
 package com.pathdb.pathIndex;
 
+import com.pathdb.pathIndex.diskBased.TreeNodeIDManager;
 import org.junit.Test;
-import com.pathdb.pathIndex.tree.TreeNodeIDManager;
 
 public class TreeNodeIDManagerTest
 {
